+++ src/main/java/com/dms/entitiy/CmMaterialClassMEntity.java
@@ -1,6 +1,6 @@
 package com.dms.entitiy;
 
-import java.math.BigDecimal;
+
 import java.sql.Date;
 import java.util.Objects;
 import javax.persistence.Basic;
@@ -18,7 +18,7 @@ public class CmMaterialClassMEntity {
   private String materialClsId;
   private String materialClsNm;
   private String materialClsType;
-  private BigDecimal seq;
+  private int seq;
   private String plantId;
   private String description;
   private String useYn;
@@ -61,11 +61,11 @@ public class CmMaterialClassMEntity {
 
   @Basic
   @Column(name = "SEQ")
-  public BigDecimal getSeq() {
+  public int getSeq() {
     return seq;
   }
 
-  public void setSeq(BigDecimal seq) {
+  public void setSeq(int seq) {
     this.seq = seq;
   }
