+++ src/main/java/com/dms/entitiy/CmProductClassMEntity.java
@@ -1,6 +1,5 @@
 package com.dms.entitiy;
 
-import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.Objects;
 import javax.persistence.Basic;
@@ -18,7 +17,7 @@ public class CmProductClassMEntity {
   private String productClsId;
   private String productClsNm;
   private String productClsType;
-  private BigDecimal seq;
+  private int seq;
   private String plantId;
   private String description;
   private String useYn;
@@ -61,11 +60,11 @@ public class CmProductClassMEntity {
 
   @Basic
   @Column(name = "SEQ")
-  public BigDecimal getSeq() {
+  public int getSeq() {
     return seq;
   }
 
-  public void setSeq(BigDecimal seq) {
+  public void setSeq(int seq) {
     this.seq = seq;
   }
